+++ src/clients/dataSourcesClient.ts
@@ -2,7 +2,7 @@ import { Logger } from 'pino'
 import computedConstants from '../common/computedConstants'
 import createLogger from '../common/logger/factory'
 import { IPaginatedResponse } from '../common/models/paginatedResponse'
-import { DataSourceDTO } from '../services/dataSources/dto/dataSource'
+import { DataSourceDTO } from '../services/dataSourceManager/dto/dataSource'
 import axios from 'axios'
 
 export class DataSourceClient {
