+++ src/config/taskConfig.ts
@@ -2,6 +2,7 @@ import config from 'config'
 
 export interface IFetchTaskConfig {
   readonly batchSize: number
+  readonly sourceApiUrl: string
 }
 
 export default function buildConfig<T> (scope: string): T {
