+++ src/monoApp.ts
@@ -3,10 +3,10 @@
 import { fastify, FastifyInstance } from 'fastify'
 import cluster from 'cluster'
 import { cpus } from 'os'
-import { DataSourceService } from './services/dataSources/dataSourceService'
+import { DataSourceService } from './services/dataSourceManager/dataSourceService'
 import createLogger from './common/logger/factory'
 import computedConstants from './common/computedConstants'
-import { WatchManagementService } from './services/watchManagement/watchManagementService'
+import { WatchManagementService } from './services/watchManager/watchManagementService'
 import { TheatreService } from './services/theatre/theatreService'
 import { IService } from './common/interfaces/service'
 import { Primary } from './primary'
