+++ src/services/dataSourceManager/package.json
