+++ src/services/watchManager/package.json
